started javafx
Simplified ex Example_Site_API and ConnectThread
Add A controller. Basically, if well handle, could be possible to interact directly
without a Singleton of ExampleSiteAPI
(which is still also a RobotListener and from now an Application)

The remaining process at the end preventing the application from
closing properly seems to come from the Bluetooth part...
Will have to check how to get rid of it.